old
chore: project created with projen

chore: create boilerplate cdk infrastructure package

chore: create boilerplate api project

chore: update journey-jigsaw project files

chore: setup basic react app boiler plate

chore: prove that imports of apis are possible in react project

feat: implement static website with basic Google OAuth flow

the OAuth flow isn't actually giving special access to any content. All
that's working for that is just when you go through the 'sign in with
google' flow, this will create a Cognito user in our backend.

fix: store Google OAuth client secret in secrets manager

I also revoked the raw client secret, and replaced it. Note that secrets
are created manually via the Console.

docs: add setup instructions to the README

chore: updated dependencies

feat: implements oauth2 flow to completion

This commit completes the oauth2 flow, in that users can get an access
token for Cognito all through the browser. This will happen in a few
steps:
1. The user clicks the 'Login with Google' button on the home page.
2. This button is just a link which redirects them to the 'hosted ui' for the cognito app client.
3. Since we only have one IDP, Cognito is smart enough to just redirect
  to Google's page rather than displaying a menu of IDP options.
4. After google authenticates them, Google will redirect them back to
  our home page with a 'code' query parameter.
5. Whenever app.tsx is loaded with a 'code' query parameter, it will
  contact our backend (via the '/access-token' GET api) to exchange
  this code for an access token.
6. Our backend makes an authenticated request to the Cognito
  '/oauth2/token' endpoint (using the client secret for the Cognito
  app client) and the code
  given to us by the user to get an access token for the user.
7. With this access token, the user is authenticated as a particular
  Cognito user.

I made one weird decision which was to create a public endpoint where
users can get the client id. This was because in development, I was
deleting the user pool and recreating it a lot and it was bothersome to
have to make frontend changes everytime I did that. With this API, the
frontend will always know the Cognito user pool client id (so that the
button is correct).

I also deleted the table since it wasn't doing anything, but will need
to add it back later.

fix: store idToken and refreshToken as well.

feat(ui): creates a basic frontend ui using MUI and with routing

This commit introduced a few things which are necessary to get a decent
frontend. Specifically, we needed to add an attribute mapping from
Google's attributes to Cognito. Doing this will ensure the idToken's
payload actually contains a url for the profile picture and the
customer's full name. This is useful in getting the "Avatar" component
to actually display the customer's profile picture with their first name as
its alternate text.
For routing to work with react-router-dom, we need to direct the user
to the index.html even when the page is not found. This commit made a
lot of fun aesthetic improvements to, such as adding a favicon of our
new logo, which was created with the help of DALL-E.

feat(api): added apis for sharing a moment and fetching the news feed

Since these APIs are sensitive (particularly making a post), I had to create an authorizer for the user pool and so these are protecting both the '/feed' 'GET' and '/moment' 'POST' endpoints. the '/feed' 'GET' api is making use of smithy's @paginated trait, which ensured we were using an industry-standard design for api's with paginated results. We had to add a GSI to the table to ensure we were able to efficiently query chronologically.